Move shape definitions to their own file/module
This involves making the vertices and elements of VertexSpecification
public, which is not harmful.
