+++ lib/graphics/src/lib.rs
@@ -5,10 +5,11 @@ mod app;
 mod window;
 mod vertex;
 mod shader;
+mod shapes;
 
 pub use app::App;
 pub use shader::SIMPLE_VERTEX_SOURCE;
 pub use shader::SIMPLE_FRAGMENT_SOURCE;
-pub use vertex::SimpleRect;
-pub use vertex::SimpleTriangle;
+pub use shapes::SimpleRect;
+pub use shapes::SimpleTriangle;
 pub use vertex::VertexSpecable;
